+++ Docs/Changelog.md
@@ -1,4 +1,13 @@
 # Changelog 📃
 
+## V1.1.0 - 2026-07-15
+
+### Added
+- New constraint: Fuel consumption
+- Truck data updated: Add alpha an beta values for fuel consumption calculations
+
+### Changed
+- Class rename: Driver → Truck
+
 ## V1.0.0 - 2026-07-12
 First functional version